Got rid of getInstance from the strip stuff and changed all of the associated things to just use an object, changed SetAllianceColor to be a method in StripSubsystem and not a Command, Changed some light settings for which colors it uses, such as when the snek sensors are tripped
The lower and upper sensor of the now get different colors if only one of them is tripped. The lower sensor is Green(for now), the upper to yellow(for Now) and white for both.

Fixed a logic thing in the strip subsystem involving the setAllianceColor and the DriverStation object/enum

updated wpilib

Deleted the set alliance color command because it is now replaced in the subsystem

Tried to make the lights turn green when the climber is at the height (+/- 2). Also changed it to only one color if there is one ball anywhere.

quick if order change